The continuing saga of how to get PWM
finally looking at opting towards a brute-force method by manually
defining the waveform for the PWM. That is, 4 defined points per period.
In order to keep this organized, we're looking at a voltage source made
up of over 50,000 lines of code. Buckle your seatbelts and hold onto
your hats folks... it's gonna be a derpy ride