Playing with buck converters
Running a number of different simulations to address issue where the
first buck converter output voltage is higher than expected. This is
thought to be occuring due to a lower current draw in the first buck
converter than in the second buck converter which is connected to the
load Thus, the first buck converter is not fully discharging each cycle.
This does seem to go away with a higher current, as demonstrated by
using a low resistance load.

Also, added 100 ohm resistors in series with the freewheeling diodes on
the filters, per Prof. E's suggestion