Added an Integral component to the control loop
Now controls voltage with a PI loop, as this allows for the voltage to
get closer to the desired value, and helps it get there quicker.

I also added code to detect when the input voltage is zero. After
finding the input to be zero for a short period it stops the output. It
then turns the buck converter on temporarily to see if there is any
power available. This also solves an issue where the integrator would
get quite large when it shouldn't.