Trying to simulate Solar Sim
Theory and practice... Unfortunately these two things rarely allign. In
this case, the device works in practice, but not in theory! (or at least
not in simulation). Python scripts are an attempt to organize data from
PSpice so it can be coppied to Excel to make an IV curve graph